Откатил класс пикселя до uint8_t

+++ RGBCell.cpp
@@ -7,39 +7,39 @@ RGBCell::RGBCell()
     _blueColor = 0;
 };
 
-RGBCell::RGBCell(int r, int g, int b)
+RGBCell::RGBCell(uint8_t r, uint8_t g, uint8_t b)
 {
     _redColor = r;
     _greenColor = g;
     _blueColor = b;
 };
 
-int RGBCell::getRedValue()
+uint8_t RGBCell::getRedValue()
 {
     return _redColor;
 };
 
-int RGBCell::getGreenValue()
+uint8_t RGBCell::getGreenValue()
 {
     return _greenColor;
 };
 
-int RGBCell::getBlueValue()
+uint8_t RGBCell::getBlueValue()
 {
     return _blueColor;
 };
 
-void RGBCell::setRedValue(int value)
+void RGBCell::setRedValue(uint8_t value)
 {
     _redColor = value;
 };
 
-void RGBCell::setGreenValue(int value)
+void RGBCell::setGreenValue(uint8_t value)
 {
     _greenColor = value;
 };
 
-void RGBCell::setBlueValue(int value)
+void RGBCell::setBlueValue(uint8_t value)
 {
     _blueColor = value;
 };
